test: move advisortest to match flat *test convention, revive dead servertest e2e files
- Move internal/testsupport/advisortest -> internal/advisortest: the repo's
  existing shared test-helper packages (protocoltest, servertest, testguard,
  vmodeltest) are flat *test packages; don't introduce a testsupport/ parent
  for a single member.
- servertest: the four files behind //go:build e2e (adaptor, load_balancing,
  server, lb_virtual_validation) no longer compiled - NewLoadBalancer moved
  to internal/protocolserver and server_test.go's import was left pointing at
  the wrong package, unnoticed because nothing ever builds with -tags=e2e
  except internal/protocol/transform. Fix the imports, and drop the e2e tag:
  every test in the package is local-only (httptest + TempDir), so the tag
  only served to hide them from CI and let them rot. Verified with -race and
  the statistical LB validation has ~1e-21 flake probability at its 65%
  threshold (binomial n=1000).
- servertest: rename utils.go/mock_provider.go to _test.go files; the package
  has no external importers, so the helpers do not need to be part of the
  production build graph.
