+++ internal/protocolserver/mcp_tool_error_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/tingly-dev/tingly-box/internal/protocol"
 	"github.com/tingly-dev/tingly-box/internal/protocolserver/advisortool"
 	"github.com/tingly-dev/tingly-box/internal/protocolserver/servertool"
-	"github.com/tingly-dev/tingly-box/internal/testsupport/advisortest"
+	"github.com/tingly-dev/tingly-box/internal/advisortest"
 	coretool "github.com/tingly-dev/tingly-box/internal/tool"
 	"github.com/tingly-dev/tingly-box/internal/typ"
 )
