+++ internal/servertest/lb_virtual_validation_test.go
@@ -1,6 +1,3 @@
-//go:build e2e
-// +build e2e
-
 package servertest
 
 import (
@@ -8,6 +5,7 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/require"
+	server "github.com/tingly-dev/tingly-box/internal/protocolserver"
 	"github.com/tingly-dev/tingly-box/internal/routing"
 
 	"github.com/tingly-dev/tingly-box/internal/config"
