+++ internal/servertest/load_balancing_test.go
@@ -1,6 +1,3 @@
-//go:build e2e
-// +build e2e
-
 package servertest
 
 import (
@@ -14,6 +11,7 @@ import (
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	server "github.com/tingly-dev/tingly-box/internal/protocolserver"
 	"github.com/tingly-dev/tingly-box/internal/routing"
 
 	"github.com/tingly-dev/tingly-box/internal/config"
