+++ internal/servertest/server_test.go
@@ -1,6 +1,3 @@
-//go:build e2e
-// +build e2e
-
 package servertest
 
 import (
@@ -10,7 +7,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/tingly-dev/tingly-box/internal/protocolserver"
+	server "github.com/tingly-dev/tingly-box/internal/server"
 
 	"github.com/tingly-dev/tingly-box/internal/server/config"
 	typ "github.com/tingly-dev/tingly-box/internal/typ"
